+++ filter_test.go
@@ -8,7 +8,7 @@ import (
 	. "github.com/dogmatiq/dapper"
 )
 
-func TestPrinter_Filter(t *testing.T) {
+func TestPrinter_WithFilter(t *testing.T) {
 	t.Run("it is passed a valid format function", func(t *testing.T) {
 		type testType struct {
 			i int
@@ -18,34 +18,32 @@ func TestPrinter_Filter(t *testing.T) {
 			i: 100,
 		}
 
-		p := &Printer{
-			Config: Config{
-				Filters: []Filter{
-					func(r Renderer, v Value) {
-						if v.DynamicType != reflect.TypeOf(testType{}) {
-							return
-						}
+		p := NewPrinter(
+			WithFilter(
+				func(r Renderer, v Value) {
+					if v.DynamicType != reflect.TypeOf(testType{}) {
+						return
+					}
 
-						r.Print("github.com/dogmatiq/dapper_test.testType<")
+					r.Print("github.com/dogmatiq/dapper_test.testType<")
 
-						fv := v.Value.FieldByName("i")
+					fv := v.Value.FieldByName("i")
 
-						r.WriteValue(
-							Value{
-								Value:                  fv,
-								DynamicType:            fv.Type(),
-								StaticType:             fv.Type(),
-								IsAmbiguousDynamicType: false,
-								IsAmbiguousStaticType:  false,
-								IsUnexported:           true,
-							},
-						)
+					r.WriteValue(
+						Value{
+							Value:                  fv,
+							DynamicType:            fv.Type(),
+							StaticType:             fv.Type(),
+							IsAmbiguousDynamicType: false,
+							IsAmbiguousStaticType:  false,
+							IsUnexported:           true,
+						},
+					)
 
-						r.Print(">")
-					},
+					r.Print(">")
 				},
-			},
-		}
+			),
+		)
 
 		expected := fmt.Sprintf("github.com/dogmatiq/dapper_test.testType<%d>", tt.i)
 		t.Log("expected:\n\n" + expected + "\n")